+++ src/webpack/webpack.base.config.js
@@ -1,4 +1,5 @@
 import path from 'path'
+import Utils from '../utils/'
 import moduleConfig from './module'
 import resolveConfig from './resolve'
 import externalsConfig from './externals'
@@ -6,6 +7,7 @@ import externalsConfig from './externals'
 const cwdPath = process.cwd() // 工程项目root path
 
 export default function(parentDirPath) {
+    Utils.setParentDir(parentDirPath) //设置br-bid所在目录， 方便babel-loader查找相关依赖
     //
     return {
         mode: 'production',